+++ src/Models/DatabaseUser.php
@@ -7,8 +7,12 @@ use Vdhicts\Cyberfusion\ClusterApi\Contracts\Model;
 
 class DatabaseUser implements Model
 {
+    private const DEFAULT_HOST = '%';
+
     public string $name;
-    public string $host = '';
+    public string $host = self::DEFAULT_HOST;
+    public string $password = '';
+    public string $serverSoftwareName = Database::DEFAULT_SERVER_SOFTWARE;
     public ?int $id = null;
     public ?int $clusterId = null;
     public ?string $createdAt = null;
@@ -19,7 +23,12 @@ class DatabaseUser implements Model
         $databaseUser = new self();
         $databaseUser->name = Arr::get($data, 'name');
         $databaseUser->id = Arr::get($data, 'id');
-        $databaseUser->host = Arr::get($data, 'host');
+        $databaseUser->host = Arr::get($data, 'host', self::DEFAULT_HOST);
+        $databaseUser->serverSoftwareName = Arr::get(
+            $data,
+            'server_software_name',
+            Database::DEFAULT_SERVER_SOFTWARE
+        );
         $databaseUser->clusterId = Arr::get($data, 'cluster_id');
         $databaseUser->createdAt = Arr::get($data, 'created_at');
         $databaseUser->updatedAt = Arr::get($data, 'updated_at');
@@ -31,6 +40,8 @@ class DatabaseUser implements Model
         return [
             'name' => $this->name,
             'host' => $this->host,
+            'password' => $this->password,
+            'server_software_name' => $this->serverSoftwareName,
             'id' => $this->id,
             'cluster_id' => $this->clusterId,
             'created_at' => $this->createdAt,